+++ Examples/YandexMobileAdsExample/YandexMobileAdsExample/Accessibility/YandexAdsAccessibility.swift
@@ -1,0 +1,9 @@
+enum YandexAdsAccessibility {
+    static let sticky = "yandex_sticky"
+    static let inline = "yandex_inline"
+    static let interstitial = "yandex_interstitial"
+    static let rewarded = "yandex_rewarded"
+    static let appOpenAd = "yandex_app_open_ad"
+    static let native = "yandex_native"
+    static let instream = "yandex_instream"
+}