+++ Examples/YandexMobileAdsExample/YandexMobileAdsExample/Yandex/Native/NativeTemplateViewController.swift
@@ -11,6 +11,7 @@ final class NativeTemplateViewController: UIViewController {
     private let adView: YMANativeBannerView = {
         let adView = YMANativeBannerView()
         adView.translatesAutoresizingMaskIntoConstraints = false
+        adView.accessibilityIdentifier = CommonAccessibility.bannerView
         return adView
     }()
 
@@ -23,7 +24,16 @@ final class NativeTemplateViewController: UIViewController {
         )
         button.translatesAutoresizingMaskIntoConstraints = false
         button.setTitle("Load ad", for: .normal)
+        button.accessibilityIdentifier = CommonAccessibility.loadButton
         return button
+    }()
+    
+    private lazy var stateLabel: UILabel = {
+        let label = UILabel()
+        label.accessibilityIdentifier = CommonAccessibility.stateLabel
+        label.numberOfLines = 0
+        label.translatesAutoresizingMaskIntoConstraints = false
+        return label
     }()
 
     private lazy var adLoader: YMANativeAdLoader = {
@@ -45,6 +55,7 @@ final class NativeTemplateViewController: UIViewController {
         // Replace demo-native-content-yandex with actual Ad Unit ID
         let requestConfiguration = YMANativeAdRequestConfiguration(adUnitID: "demo-native-content-yandex")
         adLoader.loadAd(with: requestConfiguration)
+        stateLabel.text = nil
     }
 
     private func bindNativeAd(_ ad: YMANativeAd) {
@@ -62,17 +73,22 @@ final class NativeTemplateViewController: UIViewController {
     private func addSubviews() {
         view.addSubview(adView)
         view.addSubview(loadButton)
+        view.addSubview(stateLabel)
     }
 
     private func setupConstraints() {
         let layoutGuide = view.safeAreaLayoutGuide
         NSLayoutConstraint.activate([
-            loadButton.topAnchor.constraint(equalTo: view.topAnchor, constant: 100),
+            loadButton.topAnchor.constraint(equalTo: view.safeAreaLayoutGuide.topAnchor, constant: 10),
             loadButton.centerXAnchor.constraint(equalTo: view.centerXAnchor),
 
             adView.leadingAnchor.constraint(equalTo: layoutGuide.leadingAnchor, constant: 10),
             adView.trailingAnchor.constraint(equalTo: layoutGuide.trailingAnchor, constant: -10),
-            adView.bottomAnchor.constraint(equalTo: layoutGuide.bottomAnchor, constant: -10)
+            adView.bottomAnchor.constraint(equalTo: layoutGuide.bottomAnchor, constant: -10),
+            
+            stateLabel.topAnchor.constraint(equalTo: loadButton.bottomAnchor, constant: 10),
+            stateLabel.leadingAnchor.constraint(equalTo: view.safeAreaLayoutGuide.leadingAnchor, constant: 10),
+            stateLabel.trailingAnchor.constraint(equalTo: view.safeAreaLayoutGuide.trailingAnchor, constant: -10),
         ])
     }
 }
@@ -83,10 +99,12 @@ extension NativeTemplateViewController: YMANativeAdLoaderDelegate {
     func nativeAdLoader(_ loader: YMANativeAdLoader, didLoad ad: YMANativeAd) {
         print(#function)
         bindNativeAd(ad)
+        stateLabel.text = StateUtils.loaded()
     }
 
     func nativeAdLoader(_ loader: YMANativeAdLoader, didFailLoadingWithError error: Error) {
         print(#function + "Error: \(error)")
+        stateLabel.text = StateUtils.loadError(error)
     }
 }