+++ Examples/YandexMobileAdsExample/YandexMobileAdsExampleUITests/Tests/YandexNativeTest.swift
@@ -1,0 +1,30 @@
+import XCTest
+
+final class YandexNativeTest: BaseTest {
+    let nativePage = YandexNativeAdPage()
+    let nativeListPage = YandexNativePage()
+    
+    func testNativeCustom() throws {
+        launchApp()
+        rootPage.openYandex()
+        yandexAdsPage.openNative()
+        nativeListPage.openCustom()
+        nativePage.tapLoadAd()
+        try assertAdLoaded(stateLabel: nativePage.stateLabel)
+        nativePage.assertAdDisplayed()
+        nativePage.tapAd()
+        assertSafariOpened()
+    }
+    
+    func testNativeTemplate() throws {
+        launchApp()
+        rootPage.openYandex()
+        yandexAdsPage.openNative()
+        nativeListPage.openTemplate()
+        nativePage.tapLoadAd()
+        try assertAdLoaded(stateLabel: nativePage.stateLabel)
+        nativePage.assertAdDisplayed()
+        nativePage.tapAd()
+        assertSafariOpened()
+    }
+}