+++ changelog/adapter/ironsource-mobileads/CHANGELOG.md
@@ -1,9 +1,19 @@
 # Change Log
+
 All notable changes to this project will be documented in this file.
+
+## Version 6.4.0.0
+
+### Updated
+
+* Added support for Yandex Mobile Ads SDK 6.3.0
+* Added support for IronSource SDK 7.6.0.0
+* Updated minimum supported IronSource SDK version to 7.6.0.0
 
 ## Version 6.2.0.0
 
 ### Updated
+
 * Added support for banner ads
 * Added support for Yandex Mobile Ads SDK 6.2.0
 * Added support for IronSource SDK 7.5.1.0
@@ -12,6 +22,7 @@ All notable changes to this project will be documented in this file.
 ## Version 6.1.0.0
 
 ### Updated
+
 * Added support for Yandex Mobile Ads SDK 6.1.0
 * Added support for IronSource SDK 7.5.0.0
 * Updated minimum supported IronSource SDK version to 7.5.0.0
@@ -19,6 +30,7 @@ All notable changes to this project will be documented in this file.
 ## Version 6.0.0.0
 
 ### Updated
+
 * Added support for Yandex Mobile Ads SDK 6.0.0
 * Added support for IronSource SDK 7.4.0.0
 * Updated minimum supported IronSource SDK version to 7.4.0.0
@@ -26,26 +38,31 @@ All notable changes to this project will be documented in this file.
 ## Version 5.9.1.0
 
 ### Updated
+
 * Added support for Yandex Mobile Ads SDK 5.9.1
 
 ## Version 5.9.0.0
 
 ### Updated
+
 * Added support for Yandex Mobile Ads SDK 5.9.0
 
 ## Version 5.8.0.0
 
 ### Updated
+
 * Added support for Yandex Mobile Ads SDK 5.8.0
 
 ## Version 5.7.0.0
 
 ### Updated
+
 * Added support for Yandex Mobile Ads SDK 5.7.0
 
 ## Version 5.6.0.0
 
 ### Updated
+
 * Added support for Yandex Mobile Ads SDK 5.6.0
 * Added support for IronSource SDK 7.3.0.0
 * Updated minimum supported IronSource SDK version to 7.3.0.0
@@ -53,6 +70,7 @@ All notable changes to this project will be documented in this file.
 ## Version 5.5.0.0
 
 ### Updated
+
 * Added support for Yandex Mobile Ads SDK 5.5.0
 * Added support for IronSource SDK 7.2.7.0
 * Updated minimum supported IronSource SDK version to 7.2.7.0
@@ -60,11 +78,13 @@ All notable changes to this project will be documented in this file.
 ## Version 5.4.0.0
 
 ### Updated
+
 * Added support for Yandex Mobile Ads SDK 5.4.0
 
 ## Version 5.3.0.0
 
 ### Updated
+
 * Added support for Yandex Mobile Ads SDK 5.3.0
 * Added support for IronSource SDK 7.2.5.1
 * Updated minimum supported IronSource SDK version to 7.2.5.1
@@ -72,14 +92,15 @@ All notable changes to this project will be documented in this file.
 ## Version 5.2.0.0
 
 ### Updated
+
 * Added support for Yandex Mobile Ads SDK 5.2.0
 * Added support for IronSource SDK 7.2.4.0
 * Updated minimum supported IronSource SDK version to 7.2.4.0
 
-
 ## Version 5.1.0.0
 
 ### Updated
+
 * Added support for Yandex Mobile Ads SDK 5.1.0
 * Added support for IronSource SDK 7.2.2.1
 * Updated minimum supported IronSource SDK version to 7.2.2.1
@@ -87,11 +108,13 @@ All notable changes to this project will be documented in this file.
 ## Version 5.0.0.1
 
 ### Updated
+
 * Added support for IronSource SDK 7.2.1.2
 * Updated minimum supported IronSource SDK version to 7.2.1
 
 ## Version 5.0.0.0
 
-#### Added
+### Added
+
 * Added Yandex interstitial mediation through IronSource
 * Added Yandex rewarded mediation through IronSource