+++ changelog/mediation/mobileads-inmobi/CHANGELOG.md
@@ -1,29 +1,41 @@
 # Change Log
+
 All notable changes to this project will be documented in this file.
+
+## Version 10.5.5.4
+
+### Added
+
+* Updated minimum supported Yandex Mobile Ads SDK version to 6.3.0
 
 ## Version 10.5.5.3
 
-#### Added
+### Added
+
 * Updated minimum supported Yandex Mobile Ads SDK version to 6.2.0
 
 ## Version 10.5.5.2
 
-#### Added
+### Added
+
 * Updated minimum supported Yandex Mobile Ads SDK version to 6.1.0
 
 ## Version 10.5.5.1
 
-#### Added
+### Added
+
 * Updated minimum supported Yandex Mobile Ads SDK version to 6.0.0
 
 ## Version 10.5.5.0
 
-#### Added
+### Added
+
 * Added support for InMobi sdk version 10.5.5
 * Updated minimum supported InMobi sdk version 10.5.5
 * Updated minimum supported Yandex Mobile Ads SDK version to 5.7.0
 
 ## Version 10.5.4.0
 
-#### Added
+### Added
+
 * Added InMobi mediation